\name{scoring.internal}
\alias{bdecomp}
\alias{betafam}
\alias{powfam}
\alias{sphfam}
\alias{logfam}
\alias{scalescores}
\alias{scoreitems}
\title{Internal scoring objects}
\description{
  Internal scoring rule objects not to be called by the user (unless you
  really want to).
}
\keyword{internal}
